Add fonts to demo app and use fonts in theme provider
