+++ react-native/components/src/theme/index.ts
@@ -1,43 +1,1 @@
-import { createTheming } from '@callstack/react-theme-provider';
-import { gray, white, blue, red } from '@pxblue/colors';
-
-export interface Theme {
-  roundness: number;
-  colors: {
-    primary: string;
-    background: string;
-    surface: string;
-    accent: string;
-    error: string;
-    text: string;
-  };
-  sizes: {
-    small: number;
-    medium: number;
-    large: number;
-    extraLarge: number;
-  };
-};
-
-const { ThemeProvider, withTheme } = createTheming<Theme>({
-  roundness: 3,
-  colors: {
-    primary: gray[600],
-    background: white[500],
-    surface: white[100],
-    accent: blue[700],
-    error: red[500],
-    text: gray[600],
-  },
-  sizes: {
-    small: 10,
-    medium: 12,
-    large: 16,
-    extraLarge: 24
-  }
-});
-
-export {
-  ThemeProvider,
-  withTheme
-};
+export { Theme, ThemeProvider, withTheme } from './theme';
